backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "testportal-backend"
version = "0.1.0"
description = "TestPortal Backend for API Testing"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "httpx",
    "python-dotenv",
    "sqlalchemy",
    "pydantic",
    "python-gitlab"
]

[tool.pytest.ini_options]
pythonpath = ["."] 